Move public status-related API types into apiserver/params/status.go.
