+++ apiserver/client/status.go
@@ -13,7 +13,6 @@ import (
 	"gopkg.in/juju/charm.v5"
 	"gopkg.in/juju/charm.v5/hooks"
 
-	"github.com/juju/juju/api"
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/constraints"
 	"github.com/juju/juju/network"
@@ -22,10 +21,10 @@ import (
 	"github.com/juju/juju/worker/uniter/operation"
 )
 
-func agentStatusFromStatusInfo(s []state.StatusInfo, kind params.HistoryKind) []api.AgentStatus {
-	result := []api.AgentStatus{}
+func agentStatusFromStatusInfo(s []state.StatusInfo, kind params.HistoryKind) []params.AgentStatus {
+	result := []params.AgentStatus{}
 	for _, v := range s {
-		result = append(result, api.AgentStatus{
+		result = append(result, params.AgentStatus{
 			Status: params.Status(v.Status),
 			Info:   v.Message,
 			Data:   v.Data,
@@ -37,7 +36,7 @@ func agentStatusFromStatusInfo(s []state.StatusInfo, kind params.HistoryKind) []
 
 }
 
-type sortableStatuses []api.AgentStatus
+type sortableStatuses []params.AgentStatus
 
 func (s sortableStatuses) Len() int {
 	return len(s)
@@ -51,25 +50,25 @@ func (s sortableStatuses) Less(i, j int) bool {
 
 // TODO(perrito666) this client method requires more testing, only its parts are unittested.
 // UnitStatusHistory returns a slice of past statuses for a given unit.
-func (c *Client) UnitStatusHistory(args params.StatusHistory) (api.UnitStatusHistory, error) {
+func (c *Client) UnitStatusHistory(args params.StatusHistory) (params.UnitStatusHistory, error) {
 	size := args.Size - 1
 	if size < 1 {
-		return api.UnitStatusHistory{}, errors.Errorf("invalid history size: %d", args.Size)
+		return params.UnitStatusHistory{}, errors.Errorf("invalid history size: %d", args.Size)
 	}
 	unit, err := c.api.state.Unit(args.Name)
 	if err != nil {
-		return api.UnitStatusHistory{}, errors.Trace(err)
+		return params.UnitStatusHistory{}, errors.Trace(err)
 	}
-	statuses := api.UnitStatusHistory{}
+	statuses := params.UnitStatusHistory{}
 	if args.Kind == params.KindCombined || args.Kind == params.KindWorkload {
 		unitStatuses, err := unit.StatusHistory(size)
 		if err != nil {
-			return api.UnitStatusHistory{}, errors.Trace(err)
+			return params.UnitStatusHistory{}, errors.Trace(err)
 		}
 
 		current, err := unit.Status()
 		if err != nil {
-			return api.UnitStatusHistory{}, errors.Trace(err)
+			return params.UnitStatusHistory{}, errors.Trace(err)
 		}
 		unitStatuses = append(unitStatuses, current)
 
@@ -79,12 +78,12 @@ func (c *Client) UnitStatusHistory(args params.StatusHistory) (api.UnitStatusHis
 		agent := unit.Agent()
 		agentStatuses, err := agent.StatusHistory(size)
 		if err != nil {
-			return api.UnitStatusHistory{}, errors.Trace(err)
+			return params.UnitStatusHistory{}, errors.Trace(err)
 		}
 
 		current, err := agent.Status()
 		if err != nil {
-			return api.UnitStatusHistory{}, errors.Trace(err)
+			return params.UnitStatusHistory{}, errors.Trace(err)
 		}
 		agentStatuses = append(agentStatuses, current)
 
@@ -103,12 +102,12 @@ func (c *Client) UnitStatusHistory(args params.StatusHistory) (api.UnitStatusHis
 }
 
 // FullStatus gives the information needed for juju status over the api
-func (c *Client) FullStatus(args params.StatusParams) (api.Status, error) {
+func (c *Client) FullStatus(args params.StatusParams) (params.FullStatus, error) {
 	cfg, err := c.api.state.EnvironConfig()
 	if err != nil {
-		return api.Status{}, errors.Annotate(err, "could not get environ config")
+		return params.FullStatus{}, errors.Annotate(err, "could not get environ config")
 	}
-	var noStatus api.Status
+	var noStatus params.FullStatus
 	var context statusContext
 	if context.services, context.units, context.latestCharms, err =
 		fetchAllServicesAndUnits(c.api.state, len(args.Patterns) <= 0); err != nil {
@@ -194,7 +193,7 @@ func (c *Client) FullStatus(args params.StatusParams) (api.Status, error) {
 		}
 	}
 
-	return api.Status{
+	return params.FullStatus{
 		EnvironmentName: cfg.Name(),
 		Machines:        processMachines(context.machines),
 		Services:        context.processServices(),
@@ -204,16 +203,16 @@ func (c *Client) FullStatus(args params.StatusParams) (api.Status, error) {
 }
 
 // Status is a stub version of FullStatus that was introduced in 1.16
-func (c *Client) Status() (api.LegacyStatus, error) {
-	var legacyStatus api.LegacyStatus
+func (c *Client) Status() (params.LegacyStatus, error) {
+	var legacyStatus params.LegacyStatus
 	status, err := c.FullStatus(params.StatusParams{})
 	if err != nil {
 		return legacyStatus, err
 	}
 
-	legacyStatus.Machines = make(map[string]api.LegacyMachineStatus)
+	legacyStatus.Machines = make(map[string]params.LegacyMachineStatus)
 	for machineName, machineStatus := range status.Machines {
-		legacyStatus.Machines[machineName] = api.LegacyMachineStatus{
+		legacyStatus.Machines[machineName] = params.LegacyMachineStatus{
 			InstanceId: string(machineStatus.InstanceId),
 		}
 	}
@@ -377,9 +376,9 @@ func (m machineAndContainers) Containers(id string) []*state.Machine {
 	return m[id][1:]
 }
 
-func processMachines(idToMachines map[string][]*state.Machine) map[string]api.MachineStatus {
-	machinesMap := make(map[string]api.MachineStatus)
-	cache := make(map[string]api.MachineStatus)
+func processMachines(idToMachines map[string][]*state.Machine) map[string]params.MachineStatus {
+	machinesMap := make(map[string]params.MachineStatus)
+	cache := make(map[string]params.MachineStatus)
 	for id, machines := range idToMachines {
 
 		if len(machines) <= 0 {
@@ -405,7 +404,7 @@ func processMachines(idToMachines map[string][]*state.Machine) map[string]api.Ma
 	return machinesMap
 }
 
-func makeMachineStatus(machine *state.Machine) (status api.MachineStatus) {
+func makeMachineStatus(machine *state.Machine) (status params.MachineStatus) {
 	status.Id = machine.Id()
 	agentStatus, compatStatus := processMachine(machine)
 	status.Agent = agentStatus
@@ -449,19 +448,19 @@ func makeMachineStatus(machine *state.Machine) (status api.MachineStatus) {
 	} else {
 		status.Hardware = hc.String()
 	}
-	status.Containers = make(map[string]api.MachineStatus)
+	status.Containers = make(map[string]params.MachineStatus)
 	return
 }
 
-func (context *statusContext) processRelations() []api.RelationStatus {
-	var out []api.RelationStatus
+func (context *statusContext) processRelations() []params.RelationStatus {
+	var out []params.RelationStatus
 	relations := context.getAllRelations()
 	for _, relation := range relations {
-		var eps []api.EndpointStatus
+		var eps []params.EndpointStatus
 		var scope charm.RelationScope
 		var relationInterface string
 		for _, ep := range relation.Endpoints() {
-			eps = append(eps, api.EndpointStatus{
+			eps = append(eps, params.EndpointStatus{
 				ServiceName: ep.ServiceName,
 				Name:        ep.Name,
 				Role:        ep.Role,
@@ -471,7 +470,7 @@ func (context *statusContext) processRelations() []api.RelationStatus {
 			relationInterface = ep.Interface
 			scope = ep.Scope
 		}
-		relStatus := api.RelationStatus{
+		relStatus := params.RelationStatus{
 			Id:        relation.Id(),
 			Key:       relation.String(),
 			Interface: relationInterface,
@@ -499,16 +498,16 @@ func (context *statusContext) getAllRelations() []*state.Relation {
 	return out
 }
 
-func (context *statusContext) processNetworks() map[string]api.NetworkStatus {
-	networksMap := make(map[string]api.NetworkStatus)
+func (context *statusContext) processNetworks() map[string]params.NetworkStatus {
+	networksMap := make(map[string]params.NetworkStatus)
 	for name, network := range context.networks {
 		networksMap[name] = context.makeNetworkStatus(network)
 	}
 	return networksMap
 }
 
-func (context *statusContext) makeNetworkStatus(network *state.Network) api.NetworkStatus {
-	return api.NetworkStatus{
+func (context *statusContext) makeNetworkStatus(network *state.Network) params.NetworkStatus {
+	return params.NetworkStatus{
 		ProviderId: network.ProviderId(),
 		CIDR:       network.CIDR(),
 		VLANTag:    network.VLANTag(),
@@ -536,15 +535,15 @@ func paramsJobsFromJobs(jobs []state.MachineJob) []multiwatcher.MachineJob {
 	return paramsJobs
 }
 
-func (context *statusContext) processServices() map[string]api.ServiceStatus {
-	servicesMap := make(map[string]api.ServiceStatus)
+func (context *statusContext) processServices() map[string]params.ServiceStatus {
+	servicesMap := make(map[string]params.ServiceStatus)
 	for _, s := range context.services {
 		servicesMap[s.Name()] = context.processService(s)
 	}
 	return servicesMap
 }
 
-func (context *statusContext) processService(service *state.Service) (status api.ServiceStatus) {
+func (context *statusContext) processService(service *state.Service) (status params.ServiceStatus) {
 	serviceCharmURL, _ := service.CharmURL()
 	status.Charm = serviceCharmURL.String()
 	status.Exposed = service.IsExposed()
@@ -579,7 +578,7 @@ func (context *statusContext) processService(service *state.Service) (status api
 		// <svc> --networks=...") will be enabled, and altough when
 		// specified, networks constraints will be used for instance
 		// selection, they won't be actually enabled.
-		status.Networks = api.NetworksSpecification{
+		status.Networks = params.NetworksSpecification{
 			Enabled:  networks,
 			Disabled: append(cons.IncludeNetworks(), cons.ExcludeNetworks()...),
 		}
@@ -605,15 +604,15 @@ func isColorStatus(code state.MeterStatusCode) bool {
 	return code == state.MeterGreen || code == state.MeterAmber || code == state.MeterRed
 }
 
-func (context *statusContext) processUnitMeterStatuses(units map[string]*state.Unit) map[string]api.MeterStatus {
-	unitsMap := make(map[string]api.MeterStatus)
+func (context *statusContext) processUnitMeterStatuses(units map[string]*state.Unit) map[string]params.MeterStatus {
+	unitsMap := make(map[string]params.MeterStatus)
 	for _, unit := range units {
 		status, err := unit.GetMeterStatus()
 		if err != nil {
 			continue
 		}
 		if isColorStatus(status.Code) {
-			unitsMap[unit.Name()] = api.MeterStatus{Color: strings.ToLower(status.Code.String()), Message: status.Info}
+			unitsMap[unit.Name()] = params.MeterStatus{Color: strings.ToLower(status.Code.String()), Message: status.Info}
 		}
 	}
 	if len(unitsMap) > 0 {
@@ -622,16 +621,16 @@ func (context *statusContext) processUnitMeterStatuses(units map[string]*state.U
 	return nil
 }
 
-func (context *statusContext) processUnits(units map[string]*state.Unit, serviceCharm string) map[string]api.UnitStatus {
-	unitsMap := make(map[string]api.UnitStatus)
+func (context *statusContext) processUnits(units map[string]*state.Unit, serviceCharm string) map[string]params.UnitStatus {
+	unitsMap := make(map[string]params.UnitStatus)
 	for _, unit := range units {
 		unitsMap[unit.Name()] = context.processUnit(unit, serviceCharm)
 	}
 	return unitsMap
 }
 
-func (context *statusContext) processUnit(unit *state.Unit, serviceCharm string) api.UnitStatus {
-	var result api.UnitStatus
+func (context *statusContext) processUnit(unit *state.Unit, serviceCharm string) params.UnitStatus {
+	var result params.UnitStatus
 	result.PublicAddress, _ = unit.PublicAddress()
 	unitPorts, _ := unit.OpenedPorts()
 	for _, port := range unitPorts {
@@ -647,7 +646,7 @@ func (context *statusContext) processUnit(unit *state.Unit, serviceCharm string)
 	processUnitAndAgentStatus(unit, &result)
 
 	if subUnits := unit.SubordinateNames(); len(subUnits) > 0 {
-		result.Subordinates = make(map[string]api.UnitStatus)
+		result.Subordinates = make(map[string]params.UnitStatus)
 		for _, name := range subUnits {
 			subUnit := context.unitByName(name)
 			// subUnit may be nil if subordinate was filtered out.
@@ -697,7 +696,7 @@ type lifer interface {
 }
 
 // processUnitAndAgentStatus retrieves status information for both unit and unitAgents.
-func processUnitAndAgentStatus(unit *state.Unit, status *api.UnitStatus) {
+func processUnitAndAgentStatus(unit *state.Unit, status *params.UnitStatus) {
 	status.UnitAgent, status.Workload = processUnitStatus(unit)
 
 	// Legacy fields required until Juju 2.0.
@@ -727,7 +726,7 @@ func processUnitAndAgentStatus(unit *state.Unit, status *api.UnitStatus) {
 }
 
 // populateStatusFromGetter creates status information for machines, units.
-func populateStatusFromGetter(agent *api.AgentStatus, getter state.StatusGetter) {
+func populateStatusFromGetter(agent *params.AgentStatus, getter state.StatusGetter) {
 	statusInfo, err := getter.Status()
 	agent.Err = err
 	agent.Status = params.Status(statusInfo.Status)
@@ -738,7 +737,7 @@ func populateStatusFromGetter(agent *api.AgentStatus, getter state.StatusGetter)
 
 // processMachine retrieves version and status information for the given machine.
 // It also returns deprecated legacy status information.
-func processMachine(machine *state.Machine) (out api.AgentStatus, compat api.AgentStatus) {
+func processMachine(machine *state.Machine) (out params.AgentStatus, compat params.AgentStatus) {
 	out.Life = processLife(machine)
 
 	if t, err := machine.AgentTools(); err == nil {
@@ -789,7 +788,7 @@ func processMachine(machine *state.Machine) (out api.AgentStatus, compat api.Age
 }
 
 // processUnit retrieves version and status information for the given unit.
-func processUnitStatus(unit *state.Unit) (agentStatus, workloadStatus api.AgentStatus) {
+func processUnitStatus(unit *state.Unit) (agentStatus, workloadStatus params.AgentStatus) {
 	// First determine the agent status information.
 	unitAgent := unit.Agent()
 	populateStatusFromGetter(&agentStatus, unitAgent)
@@ -803,7 +802,7 @@ func processUnitStatus(unit *state.Unit) (agentStatus, workloadStatus api.AgentS
 	return
 }
 
-func canBeLost(status *api.UnitStatus) bool {
+func canBeLost(status *params.UnitStatus) bool {
 	// Pending and Installing are deprecated.
 	// Need to still check pending for existing deployments.
 	switch status.UnitAgent.Status {
@@ -822,7 +821,7 @@ func canBeLost(status *api.UnitStatus) bool {
 // processUnitLost determines whether the given unit should be marked as lost.
 // TODO(fwereade/wallyworld): this is also model-level code and should sit in
 // between state and this package.
-func processUnitLost(unit *state.Unit, status *api.UnitStatus) {
+func processUnitLost(unit *state.Unit, status *params.UnitStatus) {
 	if !canBeLost(status) {
 		// The status is allocating or installing - there's no point
 		// in enquiring about the agent liveness.
